refactor: Centralize configurable items in a single file.
This way, we can load a configuration file in the main document, which contains the content and fields that need to be queried, as well as possible sorting methods, making it easy to quickly switch working objects.

For example, here we simply switched to another configuration file (unfortunately, it can only rely on hard-coded dict instead of external json to obtain information), and immediately started querying all the airports within Huaxia.