fix: Forgot to write the publish command.
Yes, you read it correctly. The "publish" command of the hosting service currently being used is that simple.

If you are using GitHub Pages or Vercel, it may be a bit more complicated. However, for hosting services that can deploy tar/zip packages, they should all be compatible with the current build framework.